move getTestWorkspaceFolder to workspace module
