+++ packages/salesforcedx-vscode-lwc/src/testSupport/workspace/index.ts
@@ -6,6 +6,12 @@
  */
 import { getCliArgsFromJestArgs } from './getCliArgsFromJestArgs';
 import { getLwcTestRunnerExecutable } from './getLwcTestRunnerExecutable';
+import { getTestWorkspaceFolder } from './getTestWorkspaceFolder';
 import { workspaceService } from './workspaceService';
 
-export { getCliArgsFromJestArgs, getLwcTestRunnerExecutable, workspaceService };
+export {
+  getCliArgsFromJestArgs,
+  getLwcTestRunnerExecutable,
+  getTestWorkspaceFolder,
+  workspaceService
+};
